requirements.txt
pyvcf == 0.6.8
scipy == 1.5.2
biopython == 1.78